requirements.txt
Pillow
reportlab
PyPDF2
reportlab
fastapi
uvicorn